realloc is up and running

+++ src/pf_realloc.c
@@ -3,9 +3,9 @@
 void	*ft_realloc(void *, size_t);
 
 int	pf_realloc(t_pf *pf, int offset) {
-	if (pf->len + offset < (long)pf->buf_sz)
+	if (pf->len + offset < pf->buf_sz)
 		return (0);
-	while (pf->len + offset > (long)pf->buf_sz)
+	while (pf->len + offset > pf->buf_sz)
 		pf->buf_sz *= 2;
 	pf->buf = ft_realloc(pf->buf, pf->buf_sz);
 	if (!pf->buf)
